Remove duplication of test code.
