+++ src/test.rs
@@ -25,10 +25,10 @@ use crate::{
     },
     commons::{
         api::{
-            AddChildRequest, CertAuthInfo, CertAuthInit, CertifiedKeyInfo, ChildHandle, Handle, ParentCaContact,
-            ParentCaReq, ParentHandle, ParentStatuses, PublicationServerUris, PublisherDetails, PublisherHandle,
-            PublisherList, RepositoryContact, ResourceClassName, ResourceSet, RoaDefinition, RoaDefinitionUpdates,
-            RtaList, RtaName, RtaPrepResponse, TypedPrefix, UpdateChildRequest,
+            AddChildRequest, CertAuthInfo, CertAuthInit, CertifiedKeyInfo, ChildHandle, Handle, ObjectName,
+            ParentCaContact, ParentCaReq, ParentHandle, ParentStatuses, PublicationServerUris, PublisherDetails,
+            PublisherHandle, PublisherList, RepositoryContact, ResourceClassKeysInfo, ResourceClassName, ResourceSet,
+            RoaDefinition, RoaDefinitionUpdates, RtaList, RtaName, RtaPrepResponse, TypedPrefix, UpdateChildRequest,
         },
         bgp::{Announcement, BgpAnalysisReport, BgpAnalysisSuggestion},
         crypto::SignSupport,
@@ -196,7 +196,7 @@ pub async fn krill_admin_expect_error(command: Command) -> Error {
     }
 }
 
-async fn refresh_all() {
+pub async fn refresh_all() {
     krill_admin(Command::Bulk(BulkCaCommand::Refresh)).await;
 }
 
@@ -615,6 +615,198 @@ pub fn definition(s: &str) -> RoaDefinition {
 
 pub fn typed_prefix(s: &str) -> TypedPrefix {
     TypedPrefix::from_str(s).unwrap()
+}
+
+pub async fn repo_update(ca: &Handle, contact: RepositoryContact) {
+    let command = Command::CertAuth(CaCommand::RepoUpdate(ca.clone(), contact));
+    krill_admin(command).await;
+}
+
+pub async fn embedded_repository_response(publisher: &PublisherHandle) -> rfc8183::RepositoryResponse {
+    let command = PubServerCommand::RepositoryResponse(publisher.clone());
+    match krill_embedded_pubd_admin(command).await {
+        ApiResponse::Rfc8183RepositoryResponse(response) => response,
+        _ => panic!("Expected repository response."),
+    }
+}
+
+pub async fn embedded_repo_add_publisher(req: rfc8183::PublisherRequest) {
+    let command = PubServerCommand::AddPublisher(req);
+    krill_embedded_pubd_admin(command).await;
+}
+
+pub async fn dedicated_repository_response(publisher: &PublisherHandle) -> rfc8183::RepositoryResponse {
+    let command = PubServerCommand::RepositoryResponse(publisher.clone());
+    match krill_dedicated_pubd_admin(command).await {
+        ApiResponse::Rfc8183RepositoryResponse(response) => response,
+        _ => panic!("Expected repository response."),
+    }
+}
+
+pub async fn dedicated_repo_add_publisher(req: rfc8183::PublisherRequest) {
+    let command = PubServerCommand::AddPublisher(req);
+    krill_dedicated_pubd_admin(command).await;
+}
+
+pub async fn set_up_ca_with_repo(ca: &Handle) {
+    init_ca(ca).await;
+
+    // Add the CA as a publisher
+    let publisher_request = publisher_request(ca).await;
+    embedded_repo_add_publisher(publisher_request).await;
+
+    // Get a Repository Response for the CA
+    let response = embedded_repository_response(ca).await;
+
+    // Update the repo for the child
+    let contact = RepositoryContact::new(response);
+    repo_update(ca, contact).await;
+}
+
+pub async fn expected_mft_and_crl(ca: &Handle, rcn: &ResourceClassName) -> Vec<String> {
+    let rc_key = ca_key_for_rcn(ca, rcn).await;
+    let mft_file = rc_key.incoming_cert().mft_name().to_string();
+    let crl_file = rc_key.incoming_cert().crl_name().to_string();
+    vec![mft_file, crl_file]
+}
+
+pub async fn expected_new_key_mft_and_crl(ca: &Handle, rcn: &ResourceClassName) -> Vec<String> {
+    let rc_key = ca_new_key_for_rcn(ca, rcn).await;
+    let mft_file = rc_key.incoming_cert().mft_name().to_string();
+    let crl_file = rc_key.incoming_cert().crl_name().to_string();
+    vec![mft_file, crl_file]
+}
+
+pub async fn expected_issued_cer(ca: &Handle, rcn: &ResourceClassName) -> String {
+    let rc_key = ca_key_for_rcn(ca, rcn).await;
+    ObjectName::from(rc_key.incoming_cert().cert()).to_string()
+}
+
+pub async fn will_publish_embedded(test_msg: &str, publisher: &PublisherHandle, files: &[String]) -> bool {
+    will_publish(test_msg, publisher, files, PubServer::Embedded).await
+}
+
+pub async fn will_publish_dedicated(test_msg: &str, publisher: &PublisherHandle, files: &[String]) -> bool {
+    will_publish(test_msg, publisher, files, PubServer::Dedicated).await
+}
+
+enum PubServer {
+    Embedded,
+    Dedicated,
+}
+
+async fn will_publish(test_msg: &str, publisher: &PublisherHandle, files: &[String], server: PubServer) -> bool {
+    let objects: Vec<_> = files.iter().map(|s| s.as_str()).collect();
+    for _ in 0..6000 {
+        let details = {
+            match &server {
+                PubServer::Dedicated => dedicated_repo_publisher_details(publisher).await,
+                PubServer::Embedded => publisher_details(publisher).await,
+            }
+        };
+
+        let current_files = details.current_files();
+
+        if current_files.len() == objects.len() {
+            let current_files: Vec<&uri::Rsync> = current_files.iter().map(|p| p.uri()).collect();
+            let mut all_matched = true;
+            for o in &objects {
+                if !current_files.iter().any(|uri| uri.ends_with(o)) {
+                    all_matched = false;
+                }
+            }
+            if all_matched {
+                return true;
+            }
+        }
+
+        sleep(Duration::from_millis(100)).await;
+    }
+
+    let details = publisher_details(publisher).await;
+
+    eprintln!(
+        "Did not find match for test: {}, for publisher: {}",
+        test_msg, publisher
+    );
+    eprintln!("Found:");
+    for file in details.current_files() {
+        eprintln!("  {}", file.uri());
+    }
+    eprintln!("Expected:");
+    for file in objects {
+        eprintln!("  {}", file);
+    }
+
+    false
+}
+
+pub async fn set_up_ca_under_parent_with_resources(ca: &Handle, parent: &ParentHandle, resources: &ResourceSet) {
+    let child_request = request(ca).await;
+    let parent = {
+        let contact = add_child_rfc6492(parent, ca, child_request, resources.clone()).await;
+        ParentCaReq::new(parent.clone(), contact)
+    };
+    add_parent_to_ca(ca, parent).await;
+    assert!(ca_contains_resources(ca, resources).await);
+}
+
+pub async fn ca_roll_init(handle: &Handle) {
+    krill_admin(Command::CertAuth(CaCommand::KeyRollInit(handle.clone()))).await;
+}
+
+pub async fn ca_roll_activate(handle: &Handle) {
+    krill_admin(Command::CertAuth(CaCommand::KeyRollActivate(handle.clone()))).await;
+}
+
+pub async fn state_becomes_new_key(handle: &Handle) -> bool {
+    for _ in 0..30_u8 {
+        let ca = ca_details(handle).await;
+
+        // wait for ALL RCs to become state new key
+        let rc_map = ca.resource_classes();
+
+        let expected = rc_map.len();
+        let mut found = 0;
+
+        for rc in rc_map.values() {
+            if let ResourceClassKeysInfo::RollNew(_) = rc.keys() {
+                found += 1;
+            }
+        }
+
+        if found == expected {
+            return true;
+        }
+
+        sleep(Duration::from_secs(1)).await
+    }
+    false
+}
+
+pub async fn state_becomes_active(handle: &Handle) -> bool {
+    for _ in 0..300 {
+        let ca = ca_details(handle).await;
+
+        // wait for ALL RCs to become state active key
+        let rc_map = ca.resource_classes();
+
+        let expected = rc_map.len();
+        let mut found = 0;
+
+        for rc in rc_map.values() {
+            if let ResourceClassKeysInfo::Active(_) = rc.keys() {
+                found += 1;
+            }
+        }
+
+        if found == expected {
+            return true;
+        }
+
+        sleep(Duration::from_millis(100)).await
+    }
+    false
 }
 
 #[cfg(test)]
